requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.1
astroid==2.5.2
async-timeout==4.0.2
asyncio-dgram==1.2.0
Brotli==1.0.9
bs4==0.0.1
cachetools==5.0.0
certifi==2020.12.5
charset-normalizer==2.0.12
chartlyrics==0.1.4
cleverwrap==0.3.0.2
click==8.1.3
colorama==0.4.4
colorful==0.5.4
decorator==5.1.1
dict.cc.py==3.1.0
dnspython==2.2.0
dnspython3==1.15.0
duckduckpy==0.2
ety==1.3.3
fake-useragent==0.1.11
forecastiopy==0.22
frozenlist==1.3.1
future==0.18.2
fuzzywuzzy==0.18.0
gazpacho==1.1
geoip2==4.6.0
greenlet==1.0.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==2.10
imgurpython==1.1.7
isort==5.8.0
joblib==1.0.1
lazy-object-proxy==1.6.0
maxminddb==2.2.0
mccabe==0.6.1
mutagen==1.45.1
mwparserfromhell==0.6.4
natsort==8.1.0
newsapi-python==0.2.6
numpy==1.23.3
oauthlib==3.1.0
opencv-python==4.6.0.66
orjson==3.5.1
Pillow==9.2.0
py-irclib==0.3.0
py-open-weather==0.1.3
pycryptodomex==3.15.0
pydantic==1.8.1
pyenchant==3.2.2
pylint==2.7.4
pymediawiki==0.7.1
pyocr==0.8.3
PySocks==1.7.1
python-Levenshtein==0.12.2
regex==2022.3.2
requests-oauthlib==1.3.0
rfc3986==1.5.0
rotten-tomatoes-scraper==1.4.0
sgmllib3k==1.0.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.2.1
thefuzz==0.19.0
toml==0.10.2
tqdm==4.60.0
treelib==1.6.1
tweepy==3.10.0
typing-extensions==3.7.4.3
Unidecode==1.3.4
validators==0.19.0
venusian==3.0.0
websockets==10.3
wrapt==1.12.1
yt-dlp==2024.3.10
google-transliteration-api==1.0.3
pydub==0.25.1
google-play-scraper==1.2.2
koo-api
revChatGPT
SQLAlchemy == 1.4.32
attrs == 23.2.0
beautifulsoup4 == 4.12.3
chardet == 5.2.0
feedparser == 6.0.11
googlemaps == 4.10.0
isodate == 0.6.1
lxml == 5.1.0
mcstatus == 11.1.1
multidict == 6.0.5
nltk == 3.8.1
psutil == 5.9.8
pyowm == 3.3.0
requests == 2.31.0
urllib3 == 2.2.1
watchdog == 4.0.0
yarl == 1.9.4
